refactor : RedisConfig 한곳에 설정
