+++ src/main/java/com/practice/OAuth2/global/config/RedisConfig.java
@@ -1,18 +1,22 @@
 package com.practice.OAuth2.global.config;
 
+import org.redisson.Redisson;
+import org.redisson.api.RedissonClient;
+import org.redisson.config.Config;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
 import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
-@EnableRedisRepositories // Redis 리포지토리 활성화
-public class RedisRepositoryConfig {
+@EnableRedisRepositories // 리프레시 토큰 저장용 - Redis 리포지토리 활성화 (CrudRepository 상속 받으면 활성화)
+public class RedisConfig {
 
     @Value("${spring.data.redis.host}")
     private String host;
@@ -20,15 +24,17 @@ public class RedisRepositoryConfig {
     @Value("${spring.data.redis.port}")
     private int port;
 
+    // Redis 연결용
     @Bean
     public RedisConnectionFactory redisConnectionFactory() {
         return new LettuceConnectionFactory(host, port);
     }
 
+    // 재발급 시 Double-checked lock 에서 Result Cashing 용
     @Bean
-    public RedisTemplate<String, Object> redisTemplate() {
+    public RedisTemplate<String, Object> redisTemplate(RedisConnectionFactory redisConnectionFactory) {
         RedisTemplate<String, Object> redisTemplate = new RedisTemplate<>();
-        redisTemplate.setConnectionFactory(redisConnectionFactory());
+        redisTemplate.setConnectionFactory(redisConnectionFactory);
 
         // Key, Value 직렬화 설정 (문자열 위주로 저장하므로 StringSerializer 사용)
         redisTemplate.setKeySerializer(new StringRedisSerializer());
@@ -36,5 +42,19 @@ public class RedisRepositoryConfig {
         redisTemplate.setValueSerializer(new GenericJackson2JsonRedisSerializer());
 
         return redisTemplate;
+    }
+
+    // 재발급 시 동시성 제어를 위한 Redisson lock 발급용
+    @Bean
+    public RedissonClient redissonClient() {
+        Config config = new Config();
+
+        // Redis 주소 앞에 "redis://"
+        String address = "redis://" + host + ":" + port;
+
+        config.useSingleServer()
+                .setAddress(address);
+
+        return Redisson.create(config);
     }
 }
